Filtros - Gestion de producto Mejoras de autenticación de sesión Notificaciones de sesión Correcciones de estilos

+++ src/http/axios.config.ts
@@ -1,7 +1,10 @@
 import axios from "axios";
 import { validateTokenHttp } from "./authHttp";
 import { navigateTo } from "../routes/navigation";
-
+import { CustomSwal } from "../components/UI/CustomSwal/CustomSwal";
+import { usuarioStore } from "../store/usuarioStore";
+import { useLocation } from "react-router-dom";
+const location = useLocation()
 const api = axios.create({
   baseURL: import.meta.env.VITE_API_URL,
 });
@@ -13,10 +16,13 @@ api.interceptors.request.use(async (config) => {
     if (esValido) {
       config.headers.Authorization = `Bearer ${token}`;
     } else {
-      console.warn("Token inválido o usuario no logueado");
-      localStorage.clear();
-      navigateTo("/register");
-      throw new axios.Cancel("Token inválido o usuario no logueado");
+      navigateTo("/home");
+      if (location.pathname.startsWith('/home')) {
+        localStorage.clear();
+        usuarioStore.setState({ usuarioLogeado: null })
+        CustomSwal.fire('Error', 'La sesión ha expirado. Inicie sesión nuevamente', 'warning')
+        throw new axios.Cancel("Token inválido o usuario no logueado");
+      }
     }
   }
